cc26x0: Refactored code for sleep management into PeripheralManager
