+++ chips/cc26x0/src/lib.rs
@@ -28,6 +28,7 @@ pub mod gpio;
 
 pub mod power;
 pub mod power_manager;
+pub mod peripheral_manager;
 
 // Since the setup code is converted from C -> Rust, we
 // ignore side effects from the conversion (unused vars & muts).
